api/requirements.txt
Django==3.1.13
mysqlclient==2.0.3
django-allauth==0.44.0
django-crispy-forms==1.11.1
gunicorn==20.0.4
python-slugify==4.0.1
google-cloud-storage==1.36.0
djangorestframework==3.12.2
markdown==3.3.3
django-filter==2.4.0
django-rest-auth==0.9.5
psycopg2==2.9.3
